backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-contrib-python
pillow
